Returns more statistics once a website has been successfully pinged.
The gathered statistics include:

Address: The IP address/ or hostname of the target host in string format. (if IP address was entered it returns the host name and visa versa)
RoundTripTime: The round-trip time for a ping request-response message, in milliseconds. (Response Time)
Data Size: The size of the data sent in the ping request, in bytes.
Replies Received: The number of successful ping replies received from the target host.

Another value is also returned and that is the "Status" if the host was able to be pinged it returns a value of "0" (True) in the Display it will be Shown as "Status: Online" if it returns a value of "1" (False) then the Ping request was unsuccessful, meaning statistics can't be gathered cause the host is offline.