Create RELEASE_NOTES.md
This app's history actually goes back to 2013 when I created the very first tool (Color Picker) in the app. 

Initially, it was just a simple Color Picker Tool for myself and close friends. I kept updating and rewriting this tool and eventually in 2017 I released it to the public as "Roselt Color Picker".

And in the back of my mind, I always knew I wanted the app to be fully cross-platform, open source, and have many many other tools in it.

And in 2020, I finally started working on "Roselt Developer Tools", one app to rule them all 😁.

Since 2017, I've kept release notes for every public release of the app and I'm proud to finally have the release notes available on GitHub too.